requirements.txt
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.3.4
attrs==19.3.0
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
black==21.7b0
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.2
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.3
django-cors-headers==3.7.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-seed==0.2.2
djangorestframework==3.11.0
djangorestframework-gis==0.17
djangorestframework-simplejwt==4.7.1
docker==4.1.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
Faker==8.4.0
fasteners==0.14.1
filelock==3.0.12
flake8==3.9.2
future==0.18.2
GDAL==3.0.4
gunicorn==20.1.0
httplib2==0.14.0
idna==2.10
importlib-metadata==1.5.0
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
mypy-extensions==0.4.3
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
PAM==0.4.2
paramiko==2.6.0
pathspec==0.9.0
pexpect==4.6.0
Pillow==8.2.0
platformdirs==2.3.0
protobuf==3.6.1
pycairo==1.16.2
pycodestyle==2.7.0
pycryptodome==3.10.1
pycups==1.9.73
pyflakes==2.3.1
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
pytz==2021.1
pyxdg==0.26
PyYAML==5.3.1
regex==2021.7.6
reportlab==3.5.34
requests==2.25.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
sqlparse==0.4.1
systemd-python==234
text-unidecode==1.3
texttable==1.6.2
tomli==1.1.0
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.7.2
wadllib==1.3.3
websocket-client==0.53.0
xkit==0.0.0
zipp==1.0.0
